refactor: refactor app.py in multiple files for readability
