+++ depictio/dash/core/__init__.py
@@ -1,0 +1,3 @@
+"""
+Core modules for the Depictio Dash application.
+"""
